+++ src/app/loadout-builder/filter/ExoticPicker.m.scss.d.ts
@@ -2,6 +2,10 @@
 // Please do not change this file!
 interface CssExports {
   'container': string;
+  'footer': string;
+  'selectedPerk': string;
+  'selectedPerks': string;
+  'submitButton': string;
 }
 export const cssExports: CssExports;
 export = cssExports;