+++ src/app/loadout-builder/filter/ExoticTile.tsx
@@ -3,6 +3,7 @@ import { TileGridTile } from 'app/dim-ui/TileGrid';
 import { t } from 'app/i18next-t';
 import { DefItemIcon } from 'app/inventory/ItemIcon';
 import { PluggableInventoryItemDefinition } from 'app/inventory/item-types';
+import { isExoticClassItemWithPerks } from 'app/inventory/store/exotic-class-item';
 import { useD2Definitions } from 'app/manifest/selectors';
 import { DestinyInventoryItemDefinition } from 'bungie-api-ts/destiny2';
 import React from 'react';
@@ -52,7 +53,6 @@ export default function ExoticTile({
 export function exoticTileInfo(defs: D2ManifestDefinitions, exotic: LockedExoticWithPlugs) {
   const { def, exoticPerk, exoticMods } = exotic;
   let perkShortDescription = exoticPerk?.displayProperties.description;
-
   if (exoticPerk) {
     for (const perk of exoticPerk.perks) {
       const description = defs.SandboxPerk.get(perk.perkHash)?.displayProperties.description;
@@ -66,7 +66,8 @@ export function exoticTileInfo(defs: D2ManifestDefinitions, exotic: LockedExotic
   const description = (
     <>
       {exotic.isArmor1 && <div>{t('LB.IncompatibleWithOptimizer')}</div>}
-      {exoticPerk && perkShortDescription}
+      {(isExoticClassItemWithPerks(exotic.def.hash) && <div>{t('LB.SelectablePerks')}</div>) ||
+        (exoticPerk && perkShortDescription)}
       {exoticMods?.map((mod) => (
         <div key={mod.hash} className={styles.perkOrModNameAndImage}>
           <DefItemIcon className={styles.perkOrModImage} itemDef={mod} />