+++ docs/CHANGELOG.md
@@ -1,5 +1,8 @@
 ## Next
 
+* Fix masterwork socket on newer armor not displaying correctly
+* Add the ability to select exotic class item perks in Loadout Optimizer
+
 ## 8.123.0 <span class="changelog-date">(2026-05-03)</span>
 
 * Restyle Collections Armor to match Universal Ornaments.